+++ backend/db/migrations/000046_add_hide_favorite_onthisday.down.sql
@@ -1,0 +1,4 @@
+ALTER TABLE devices DROP COLUMN favorites_only;
+ALTER TABLE devices DROP COLUMN on_this_day;
+ALTER TABLE images DROP COLUMN favorite;
+ALTER TABLE images DROP COLUMN hidden;